+++ packages/node/src/index.ts
@@ -1,6 +1,1 @@
-export * from './find-config'
-export * from './find-tsconfig'
-export * from './load-config'
-export * from './runtime'
-export * from './setup-config'
-export * from './setup-template'
+export * from './setup'