+++ core/dev/CHANGELOG.md
@@ -11,7 +11,7 @@
 - Updated dependencies [fde906b]
   - @rewordlabs/extractor@0.1.0
   - @rewordlabs/logger@0.1.0
-  - @rewordlabs/node@0.1.0
+  - @rewordlabs/runtime@0.1.0
   - @rewordlabs/translator@0.1.0
   - @rewordlabs/utils@0.1.0
 
@@ -25,7 +25,7 @@
 - Updated dependencies [c3b366e]
   - @rewordlabs/extractor@0.0.0
   - @rewordlabs/logger@0.0.0
-  - @rewordlabs/node@0.0.0
+  - @rewordlabs/runtime@0.0.0
   - @rewordlabs/translator@0.0.0
   - @rewordlabs/utils@0.0.0
 
@@ -39,6 +39,6 @@
 - Updated dependencies [c12aad6]
   - @rewordlabs/extractor@0.0.0-dev-20230912123137
   - @rewordlabs/logger@0.0.0-dev-20230912123137
-  - @rewordlabs/node@0.0.0-dev-20230912123137
+  - @rewordlabs/runtime@0.0.0-dev-20230912123137
   - @rewordlabs/translator@0.0.0-dev-20230912123137
   - @rewordlabs/utils@0.0.0-dev-20230912123137